requirements.txt
python-dotenv>=1.0.0
google-api-python-client
google-auth
google-auth-oauthlib
pypdf
stripe>=10.0.0
